# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ReliAPI",
    "description": "Reliability layer for API calls: retries, caching, dedup, circuit breakers, LLM proxy with cost control.",
    "version": "1.1.0",
    "contact": {
      "name": "KikuAI-Lab",
      "url": "https://github.com/kiku-jw/reliapi",
      "email": "dev@kikuai.dev"
    },
    "license": {
      "name": "AGPL-3.0-only",
      "url": "https://www.gnu.org/licenses/agpl-3.0.html"
    }
  },
  "paths": {
...
    "/onboarding/start": {
      "post": {
        "tags": [
          "onboarding"
        ],
        "summary": "Start Onboarding",
        "description": "Start self-service onboarding for a new user.\n\nThis endpoint:\n1. Generates API key automatically\n2. Creates user account\n3. Returns quick start guide and examples\n4. Provides integration verification endpoint\n\nAll steps are automated - no manual intervention required.",
        "operationId": "start_onboarding_onboarding_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnboardingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/onboarding/verify": {
      "post": {
        "tags": [
          "onboarding"
        ],
        "summary": "Verify Integration",
        "description": "Verify user integration.\n\nChecks if user has made successful API calls and provides feedback.",
        "operationId": "verify_integration_onboarding_verify_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Verify Integration Onboarding Verify Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/dashboard/api/items": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Get Items",
        "description": "\u041f\u043e\u043b\u0443\u0447\u0430\u0435\u0442 \u043e\u0442\u0444\u0438\u043b\u044c\u0442\u0440\u043e\u0432\u0430\u043d\u043d\u044b\u0435 items \u0434\u043b\u044f \u0434\u0430\u0448\u0431\u043e\u0440\u0434\u0430.",
        "operationId": "get_items_dashboard_api_items_get",
        "parameters": [
          {
            "name": "platform",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Platform"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Get Items Dashboard Api Items Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OnboardingRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "plan": {
            "type": "string",
            "title": "Plan",
            "description": "Plan to start with (free, pro, team)",
            "default": "free"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "OnboardingRequest",
        "description": "Onboarding request model."
      },
      "OnboardingResponse": {
        "properties": {
          "api_key": {
            "type": "string",
            "title": "Api Key"
          },
          "quick_start_url": {
            "type": "string",
            "title": "Quick Start Url"
          },
          "documentation_url": {
            "type": "string",
            "title": "Documentation Url"
          },
          "example_code": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Example Code"
          },
          "integration_status": {
            "type": "string",
            "title": "Integration Status"
          }
        },
        "type": "object",
        "required": [
          "api_key",
          "quick_start_url",
          "documentation_url",
          "example_code",
          "integration_status"
        ],
        "title": "OnboardingResponse",
        "description": "Onboarding response model."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}